Update repository name
The reason for the name update is that RaidLogger was too generic and not very helpful for publishing purposes. In addition to that, the subset of data being handled by the addon may expand outside of raiding purposes in the future, so there is no reason to cause confusion with the scope in the repository/addon name.